project/requirements.txt
Django==5.1.3
djangorestframework==3.15.2

# converting date to shamsi_date
jdatetime==5.0.0
